requirements.txt
PyYAML==5.4.1
selenium==3.141.0
urllib3==1.26.6
